test(html tests): restructured tests
